--- error-messages.js
@@ -1,4 +1,0 @@
-exports.IF_WITH_NO_CONDITION = '<If> tag with no condition attribute';
-exports.FOR_WITH_NO_ATTRIBUTES = '<For> tag with no \'each\' or \'of\' attribute';
-exports.MULTIPLE_CHILDREN = 'Control statements cannot have multiple children';
-exports.NO_CHILDREN = 'Control statement must have a child!';
