+++ package.json
@@ -6,7 +6,7 @@
     "name": "Alex Gilleran",
     "email": "alex@alexgilleran.com"
   },
-  "main": "index.js",
+  "main": "src/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/AlexGilleran/jsx-control-statements"
@@ -18,7 +18,9 @@
     "react",
     "jsx",
     "if",
+    "else",
     "for",
+    "each",
     "loop",
     "babel",
     "react-component"
